web/Face/Face/requirements.txt
opencv-python
pillow
numpy
mindspore
django
django-sslserver